Moved all the standard commands to its own package called stdcommands
